moved contig reader test and some test configurations to the core module. refactor
