+++ eva-accession-core/pom.xml
@@ -53,6 +53,19 @@
 
     <build>
         <finalName>${project.artifactId}-${project.version}-${git.commit.id.abbrev}</finalName>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-jar-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>test-jar</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
     </build>
 
 </project>
