+++ eva-accession-core/src/test/java/uk/ac/ebi/eva/accession/core/test/configuration/DbsnpTestDataSource.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.dbsnp.test;
+package uk.ac.ebi.eva.accession.core.test.configuration;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
