+++ eva-accession-core/src/test/java/uk/ac/ebi/eva/accession/core/test/configuration/TestConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.ac.ebi.eva.accession.release.test.configuration;
+package uk.ac.ebi.eva.accession.core.test.configuration;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
