+++ eva-accession-core/src/test/resources/contig-test.properties
@@ -1,0 +1,17 @@
+dbsnp.datasource.driver-class-name=org.hsqldb.jdbcDriver
+dbsnp.datasource.url=jdbc:hsqldb:mem:db;sql.syntax_pgs=true;DB_CLOSE_DELAY=-1
+dbsnp.datasource.username=SA
+dbsnp.datasource.password=
+dbsnp.datasource.schema=src/test/resources/test-data/dbsnp-mirror-schema.sql
+dbsnp.datasource.data=src/test/resources/test-data/dbsnp-mirror-data.sql
+
+parameters.fasta=src/test/resources/input-files/fasta/Gallus_gallus-5.0.test.fa
+parameters.assemblyReportUrl=file:src/test/resources/input-files/assembly-report/GCF_000002315.4_Gallus_gallus-5.0_assembly_report.txt
+parameters.assemblyName=Gallus_gallus-5.0
+parameters.assemblyAccession=GCF_000002315.4
+parameters.taxonomyAccession=9031
+parameters.forceRestart=false
+parameters.chunkSize=100
+parameters.pageSize=100
+
+mongodb.read-preference=primary
