+++ eva-accession-core/src/test/resources/test-data/dbsnp-mirror-data.sql
@@ -1,0 +1,11 @@
+-- batch_id, batch_name, batch_handle, tax_id, univar_id, var_str, subsnp_class, rs_id, ss_id, contig_name, contig_start, chromosome, chromosome_start, reference, alleles, loc_type, subsnp_orientation, snp_orientation, contig_orientation, asn_from, phys_pos_from, lc_ngbr, snp_validation_status, subsnp_validation_status, genotype_exists, freq_exists, ss_create_time, rs_create_time, load_order
+INSERT INTO dbsnp_variant_load_nohgvslink_d8c757988871529f37061fa9c79477a5 VALUES(11828, 'CHICKEN_SNPS_LAYER', 'BGI', 9031, 128, 'A/C/T', 1, 13677177, 26201546, 'NT_455866', 1766472, '22', 61, 'T', 'T/A', 2, 1, 1, 1, 1766471, 91223960, 1766470, 0, 1, 1, 1, '2004-06-24 12:10:00.000', '2004-07-03 16:03:00.000', 4599591);
+INSERT INTO dbsnp_variant_load_nohgvslink_d8c757988871529f37061fa9c79477a5 VALUES(1062064, 'CHICKEN_INDEL_DWBURT', 'DWBURT', 9031, 238, '-/TCGG', 2, 739617577, 1540359250, 'NT_455837', 11724980, '22', 469, 'CCGA', 'TCGG/-', 1, 1, -1, -1, 11724979, 47119826, 11724978, 0, 0, 0, 0, '2015-01-21 17:13:00.000', '2015-03-25 15:37:00.000', 7235438);
+INSERT INTO dbsnp_variant_load_nohgvslink_d8c757988871529f37061fa9c79477a5 VALUES(11825, 'CHICKEN_SNPS_BROILER', 'BGI', 9031, 125, 'A/C/G', 1, 13823349, 25062583, 'NT_464165.1', 3, NULL, 7717899, 'C', 'T/C', 2, -1, -1, 1, 7717898, 7717898, 7717897, 1, 0, 1, 1, '2004-06-22 16:27:00.000', '2004-07-02 16:03:00.000', 54215963);
+INSERT INTO dbsnp_variant_load_nohgvslink_d8c757988871529f37061fa9c79477a5 VALUES(11825, 'CHICKEN_SNPS_BROILER', 'BGI', 9031, 125, 'A/C/G', 1, 13823349, 25312601, 'AADN04000814.1', 3, NULL, 7717899, 'C', 'C/G', 2, 1, -1, 1, 7717898, 7717898, 7717897, 1, 1, 1, 0, '2004-06-22 21:41:00.000', '2004-07-02 16:03:00.000', 4545970);
+INSERT INTO dbsnp_variant_load_nohgvslink_d8c757988871529f37061fa9c79477a5 VALUES(11831, 'CHICKEN_SNPS_SILKIE', 'BGI', 9031, 125, 'A/C/G', 1, 13823349, 27587141, 'chrUn_Scaffold827', 3, NULL, 7717899, 'C', 'C/A', 2, 1, -1, 1, 7717898, 7717898, 7717897, 1, 1, 0, 1, '2004-06-25 02:52:00.000', '2004-07-02 16:03:00.000', 49852883);
+INSERT INTO dbsnp_variant_load_nohgvslink_d8c757988871529f37061fa9c79477a5 VALUES(11825, 'BATCH', 'HANDLE', 9031, 125, 'A/C/G', 1, 7777777, 88888888, 'AADN04000814.1', 4, NULL, NULL, 'C', 'C/G', 2, 1, -1, 1, 7717898, 7717898, 7717897, 1, 1, 1, 0, '2004-06-22 21:41:00.000', '2004-07-04 16:03:00.000', 88888888);
+INSERT INTO dbsnp_variant_load_nohgvslink_d8c757988871529f37061fa9c79477a5 VALUES(12345, 'BATCH', 'HANDLE', 9031, 100, '-/CG', 2, 6666666, 9999999, 'CTG1', 3, '22', 3, 'CG', 'CG/-', 2, 1, 1, 1, 3, 3, 3, 0, 0, 0, 0, '2004-06-22 21:41:00.000', '2004-07-05 16:03:00.000', 9999999);
+INSERT INTO dbsnp_nohgvs_8a503d989bf1f58e95a7861fc999ea1d_b145 VALUES(1059308, 'KOREAN_CHICKEN_G53', 'CNU_JH_AMG', 9031, 2, 'A/G', 1, 3137311, 804868313, 'NW_003763476.1', 2151199, '1', 49437283, 'A', 'A/G', 2, 1, 1, 1, 2151198, 49437282, 2151197, 3, 1, 0, 0, '2013-06-14 18:33:00.000', '2002-04-10 21:07:00.000', 5);
+INSERT INTO dbsnp_nohgvs_8a503d989bf1f58e95a7861fc999ea1d_b145 VALUES(1062063, 'CHICKEN_15M_SNP_DWBURT', 'DWBURT', 9031, 2, 'A/G', 1, 3137394, 1550081973, 'NW_003763901.1', 4214300, '12', 6953431, 'T', 'T/C', 2, -1, -1, 1, 4214299, 6953430, 4214298, 1, 1, 0, 0, '2015-01-22 06:33:00.000', '2002-04-10 21:07:00.000', 6);
+INSERT INTO dbsnp_nohgvs_8a503d989bf1f58e95a7861fc999ea1d_b145 VALUES(1062063, 'CHICKEN_15M_SNP_DWBURT', 'DWBURT', 9031, 2, 'A/G', 1, 3137594, 1542206258, 'NW_003763686.1', 14640285, '3', 115019218, 'G', 'G/A', 2, 1, 1, 1, 14640284, 115019217, 14640283, 1, 1, 0, 0, '2015-01-21 19:54:00.000', '2002-04-10 21:07:00.000', 7);
