+++ eva-accession-import/pom.xml
@@ -23,6 +23,13 @@
         </dependency>
         <dependency>
             <groupId>uk.ac.ebi.eva</groupId>
+            <artifactId>eva-accession-core</artifactId>
+            <type>test-jar</type>
+            <version>${project.version}</version>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>uk.ac.ebi.eva</groupId>
             <artifactId>variation-commons-core</artifactId>
         </dependency>
         <dependency>
