+++ eva-accession-import/src/test/java/uk/ac/ebi/eva/accession/dbsnp/io/SubSnpNoHgvsReaderTest.java
@@ -28,8 +28,8 @@ import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
 
 import uk.ac.ebi.eva.accession.core.configuration.DbsnpDataSource;
+import uk.ac.ebi.eva.accession.core.test.configuration.TestConfiguration;
 import uk.ac.ebi.eva.accession.dbsnp.model.SubSnpNoHgvs;
-import uk.ac.ebi.eva.accession.dbsnp.test.TestConfiguration;
 
 import java.util.ArrayList;
 import java.util.List;
