--- eva-accession-import/src/test/java/uk/ac/ebi/eva/accession/dbsnp/test/TestConfiguration.java
@@ -1,27 +1,0 @@
-/*
- * Copyright 2018 EMBL - European Bioinformatics Institute
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package uk.ac.ebi.eva.accession.dbsnp.test;
-
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
-
-import uk.ac.ebi.eva.accession.core.configuration.DbsnpDataSource;
-
-@Configuration
-@EnableConfigurationProperties({DbsnpDataSource.class, DbsnpTestDataSource.class})
-public class TestConfiguration {
-
-}
